+++ examples/basic_algorithms.py
@@ -2,9 +2,11 @@
 Example of usage of basic graph algorithms including BFS, etc. in Graphite with internal
 `algorithms` module.
 """
-
+import os
+import sys
 from random import randint
 
+sys.path.insert(0, os.path.abspath('..'))
 from src.graphite import Direction, GraphiteEngine
 from src.graphite.algorithms import bfs, neighborhood
 # Use this in your code instead:
@@ -40,3 +42,8 @@ print(bfs(
 print(neighborhood(engine, "node0", 4))
 # Important algorithms are available directly in engine, so this has same result:
 #print(engine.neighborhood("node0", 4))
+
+# Find nearest paths to any node from node0
+result = engine.bfs("node0", stop_at_first=False, max_results=30)
+print(result)
+print([target for target, distance, path in result])