+++ examples/example_suite.py
@@ -1,7 +1,10 @@
 """
 Example usages of Graphite in separate functions. You can run a function to see its result.
 """
+import os
+import sys
 
+sys.path.insert(0, os.path.abspath('..'))
 from src import graphite
 # Use this in your code instead:
 #import graphite